delete room functionality completed
When the owner of the room deletes the room, owner emits "deleteRoom" event to the server.

Server on getting "deleteRoom" emits the event "roomDeleted" to all other users in the room.

All other users in front-end get "roomDeleted" event and make their roomContext empty so that the users automatically disconnects and internally leave the room they have joined.

-- *Note* --
Leaving all the joined rooms on disconnecting is socket.io inbuilt function.